+++ src/__screenshot_tests__/badge-screenshot-test.js
@@ -1,0 +1,13 @@
+// @flow
+import {openStoryPage} from '../test-utils';
+
+test('Badge UI package', async () => {
+    const page = await openStoryPage({
+        section: 'Components|Hints/Badge',
+        name: 'badge',
+        device: 'MOBILE_IOS',
+    });
+
+    const image = await page.screenshot();
+    expect(image).toMatchImageSnapshot();
+});