+++ src/__stories__/header-story.js
@@ -4,10 +4,9 @@ import {MemoryRouter} from 'react-router-dom';
 import {
     Header,
     HeaderLayout,
-    MainSectionHeader,
-    MainSectionHeaderLayout,
     Box,
     ButtonPrimary,
+    ButtonSecondary,
     Stack,
     NavigationBreadcrumbs,
     ResponsiveLayout,
@@ -15,7 +14,7 @@ import {
 import {useTextField, useCheckbox} from './helpers';
 
 export default {
-    title: 'Components|Header',
+    title: 'Components|Headers/Header',
     parameters: {
         fullScreen: true,
     },
@@ -25,7 +24,7 @@ const FieldWithCheckbox = ({children}: {children: React.Node}) => (
     <div style={{display: 'flex', alignItems: 'center'}}>{children}</div>
 );
 
-export const HeaderExample = (): React.Node => {
+export const Default = (): React.Node => {
     const [pretitle, pretitleTextField] = useTextField('pretitle', 'Factura de diciembre');
     const [title, titleTextField] = useTextField(
         'title',
@@ -33,11 +32,19 @@ export const HeaderExample = (): React.Node => {
     );
     const [preamount, preamountTextField] = useTextField('preamount', 'Cuota mensual (IVA incluido)');
     const [amount, amountTextField] = useTextField('amount', '60,44 €');
-    const [buttonLabel, buttonLabelTextField] = useTextField('buttonLabel', 'Descargar factura');
+    const [buttonLabel, buttonLabelTextField] = useTextField('button', 'Descargar factura');
+    const [secondaryButtonLabel, secondaryButtonLabelTextField] = useTextField(
+        'secondaryButton',
+        'Pagar factura'
+    );
     const [subtitle, subtitleTextField] = useTextField('subtitle', 'Y esto es un subtitulo');
     const [isInverse, inverseCheckbox] = useCheckbox('Inverse', true);
     const [isErrorAmount, errorAmountCheckbox] = useCheckbox('Error amount', false);
-    const [withExtraContent, extraContentCheckbox] = useCheckbox('With secondary content', true);
+    const [withExtraContent, extraContentCheckbox] = useCheckbox('With extra content', true);
+    const [extraSideBySide, extraSideBySideCheckbox] = useCheckbox(
+        'Extra content placed on the right in desktop',
+        true
+    );
     const [withBreadcrumbs, breadcrumbsCheckbox] = useCheckbox('With breadcrumbs', true);
     return (
         <MemoryRouter>
@@ -45,6 +52,7 @@ export const HeaderExample = (): React.Node => {
                 <div data-testid="header-layout">
                     <HeaderLayout
                         isInverse={isInverse}
+                        sideBySideExtraOnDesktop={extraSideBySide}
                         breadcrumbs={
                             withBreadcrumbs ? (
                                 <NavigationBreadcrumbs
@@ -62,6 +70,11 @@ export const HeaderExample = (): React.Node => {
                                 button={
                                     buttonLabel ? (
                                         <ButtonPrimary href="asdf">{buttonLabel}</ButtonPrimary>
+                                    ) : undefined
+                                }
+                                secondaryButton={
+                                    secondaryButtonLabel ? (
+                                        <ButtonSecondary href="asdf">{secondaryButtonLabel}</ButtonSecondary>
                                     ) : undefined
                                 }
                                 subtitle={subtitle}
@@ -82,9 +95,11 @@ export const HeaderExample = (): React.Node => {
                             {errorAmountCheckbox}
                         </FieldWithCheckbox>
                         {buttonLabelTextField}
+                        {secondaryButtonLabelTextField}
                         {subtitleTextField}
                         {inverseCheckbox}
                         {extraContentCheckbox}
+                        {extraSideBySideCheckbox}
                     </Stack>
                 </ResponsiveLayout>
             </Stack>
@@ -92,33 +107,4 @@ export const HeaderExample = (): React.Node => {
     );
 };
 
-export const MainSectionHeaderExample = (): React.Node => {
-    const [title, titleTextField] = useTextField('title', 'Soporte');
-    const [desciption, descriptionTextField] = useTextField('description', '¿En qué podemos ayudarte?');
-    const [buttonLabel, buttonLabelTextField] = useTextField('buttonLabel', 'Acción');
-    const [isInverse, inverseCheckbox] = useCheckbox('Inverse', true);
-
-    return (
-        <Stack space={16}>
-            <div data-testid="header-layout">
-                <MainSectionHeaderLayout isInverse={isInverse}>
-                    <MainSectionHeader
-                        title={title}
-                        description={desciption}
-                        button={
-                            buttonLabel ? <ButtonPrimary href="asdf">{buttonLabel}</ButtonPrimary> : undefined
-                        }
-                    />
-                </MainSectionHeaderLayout>
-            </div>
-            <ResponsiveLayout>
-                <Stack space={16}>
-                    {titleTextField}
-                    {descriptionTextField}
-                    {buttonLabelTextField}
-                    {inverseCheckbox}
-                </Stack>
-            </ResponsiveLayout>
-        </Stack>
-    );
-};
+Default.story = {name: 'Header'};