+++ src/__stories__/icon-button-story.js
@@ -5,7 +5,7 @@ import IcnClose from '../icons/icn-close';
 import {StorySection} from './helpers';
 
 export default {
-    title: 'Components|IconButton',
+    title: 'Components|Touchables/IconButton',
 };
 
 export const Default = (): React.Node => (
@@ -25,3 +25,5 @@ export const Default = (): React.Node => (
         </StorySection>
     </>
 );
+
+Default.story = {name: 'IconButton'};