+++ src/theme-context-provider.js
@@ -105,6 +105,7 @@ const ThemeContextProvider = ({theme, children}: Props): React.Node => {
             divider: c.DIVIDER,
             border: c.BORDER,
             borderLight: c.BORDER_LIGHT,
+            textDanger: c.TEXT_DANGER,
             textError: c.TEXT_ERROR,
 
             feedbackErrorBackground: c.FEEDBACK_ERROR_BACKGROUND,